CLOUDP-307030 - Mck gomod (#4210)
# Summary

- remove go.mod/go.sum from mco 
  - change all imports to file instead of deps imports
- remove test image dockerfile from mco 
- create test-image under /docker/mongodb-community-tests and change all
related paths
- change all dockerfile and paths to only reference upper go.mod
- fix all linter errors, since root golangci now runs against that
code..


## Proof of Work

- ci is green, except for the known flaky tests + broken mco tests 

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise 
  * --> no prefix is considered a question

+++ mongodb-community-operator/pkg/kube/probes/probes.go
@@ -29,30 +29,33 @@ func WithExecCommand(cmd []string) Modification {
 	}
 }
 
-func WithFailureThreshold(failureThreshold int) Modification {
+func WithFailureThreshold(failureThreshold int32) Modification {
 	return func(probe *corev1.Probe) {
-		probe.FailureThreshold = int32(failureThreshold)
+		probe.FailureThreshold = failureThreshold
 	}
 }
 
-func WithInitialDelaySeconds(initialDelaySeconds int) Modification {
+func WithInitialDelaySeconds(initialDelaySeconds int32) Modification {
 	return func(probe *corev1.Probe) {
-		probe.InitialDelaySeconds = int32(initialDelaySeconds)
+		probe.InitialDelaySeconds = initialDelaySeconds
 	}
 }
-func WithSuccessThreshold(successThreshold int) Modification {
+
+func WithSuccessThreshold(successThreshold int32) Modification {
 	return func(probe *corev1.Probe) {
-		probe.SuccessThreshold = int32(successThreshold)
+		probe.SuccessThreshold = successThreshold
 	}
 }
-func WithPeriodSeconds(periodSeconds int) Modification {
+
+func WithPeriodSeconds(periodSeconds int32) Modification {
 	return func(probe *corev1.Probe) {
-		probe.PeriodSeconds = int32(periodSeconds)
+		probe.PeriodSeconds = periodSeconds
 	}
 }
-func WithTimeoutSeconds(timeoutSeconds int) Modification {
+
+func WithTimeoutSeconds(timeoutSeconds int32) Modification {
 	return func(probe *corev1.Probe) {
-		probe.TimeoutSeconds = int32(timeoutSeconds)
+		probe.TimeoutSeconds = timeoutSeconds
 	}
 }
